qol: fix anim-editor map-key truncation + silence CRT warning noise
Triaged the /W4 narrowing-conversion warnings (C4244/C4267). One real latent bug:
the anim-editor original-model cache keyed a std::map on int but the key is an
intptr_t (the character pointer), truncating it to 32 bits so two characters could
collide; widened the key to intptr_t. The rest (double->float in a float engine,
size_t->int on bounded counts, uint64->uint RNG seed) are benign.

Also define _CRT_SECURE_NO_WARNINGS on MSVC, removing ~93 C4996 "unsafe CRT"
warnings (fopen/getenv/strncpy used intentionally) so the bug-candidate warnings
stay visible under /W4.

+++ editor/src/Application.cpp
@@ -15518,7 +15518,9 @@ void Application::renderAnimEditorUI() {
             //
             // The simplest correct approach: maintain our own copy of the original model.
             // We'll use a lazy-initialized "original model" cache.
-            static std::map<int, Phyxel::VoxelModel> s_originalModels;
+            // Key by the full pointer width — an int key truncates the 64-bit pointer and
+            // two characters can collide on the low 32 bits.
+            static std::map<intptr_t, Phyxel::VoxelModel> s_originalModels;
             auto charKey = (intptr_t)m_animEditorChar;
             if (s_originalModels.find(charKey) == s_originalModels.end()) {
                 s_originalModels[charKey] = m_animEditorChar->getVoxelModel();
